compilation errors done

+++ src/movements.c
@@ -10,9 +10,9 @@ void	move_up(t_cub *c)
 	pos_x = (int)c->player->pos_x;
 	pos_y = (int)c->player->pos_y;
 	mv_speed = c->mlx->delta_time * SPEED;
-	if (c->map->m[pos_y][pos_x + c->player->dir_x * mv_speed] == FLOOR)
+	if (c->map->m[pos_y][pos_x + (int)(c->player->dir_x * mv_speed)] == FLOOR)
 		c->player->pos_x += c->player->dir_x * mv_speed;
-	if (c->map->m[pos_y - c->player->dir_y * mv_speed][pos_x] == FLOOR)
+	if (c->map->m[pos_y - (int)(c->player->dir_y * mv_speed)][pos_x] == FLOOR)
 		c->player->pos_y -= c->player->dir_y * mv_speed;
 }
 
@@ -25,9 +25,9 @@ void	move_down(t_cub *c)
 	pos_x = (int)c->player->pos_x;
 	pos_y = (int)c->player->pos_y;
 	mv_speed = c->mlx->delta_time * SPEED;
-	if (c->map->m[pos_y][pos_x - c->player->dir_x * mv_speed] == FLOOR)
+	if (c->map->m[pos_y][pos_x - (int)(c->player->dir_x * mv_speed)] == FLOOR)
 		c->player->pos_x -= c->player->dir_x * mv_speed;
-	if (c->map->m[pos_y + c->player->dir_y * mv_speed][pos_x] == FLOOR)
+	if (c->map->m[pos_y + (int)(c->player->dir_y * mv_speed)][pos_x] == FLOOR)
 		c->player->pos_y += c->player->dir_y * mv_speed;
 }
 
@@ -40,9 +40,9 @@ void	move_left(t_cub *c)
 	pos_x = (int)c->player->pos_x;
 	pos_y = (int)c->player->pos_y;
 	mv_speed = c->mlx->delta_time * SPEED;
-	if (c->map->m[pos_y][pos_x + c->player->dir_y * mv_speed] == FLOOR)
+	if (c->map->m[pos_y][pos_x + (int)(c->player->dir_y * mv_speed)] == FLOOR)
 		c->player->pos_x += c->player->dir_y * mv_speed;
-	if (c->map->m[pos_y - c->player->dir_x * mv_speed][pos_x] == FLOOR)
+	if (c->map->m[pos_y - (int)(c->player->dir_x * mv_speed)][pos_x] == FLOOR)
 		c->player->pos_y -= c->player->dir_x * mv_speed;
 }
 
@@ -55,8 +55,8 @@ void	move_right(t_cub *c)
 	pos_x = (int)c->player->pos_x;
 	pos_y = (int)c->player->pos_y;
 	mv_speed = c->mlx->delta_time * SPEED;
-	if (c->map->m[pos_y][pos_x - c->player->dir_y * mv_speed] == FLOOR)
+	if (c->map->m[pos_y][pos_x - (int)(c->player->dir_y * mv_speed)] == FLOOR)
 		c->player->pos_x += c->player->dir_y * mv_speed;
-	if (c->map->m[pos_y + c->player->dir_x * mv_speed][pos_x] == FLOOR)
+	if (c->map->m[pos_y + (int)(c->player->dir_x * mv_speed)][pos_x] == FLOOR)
 		c->player->pos_y += c->player->dir_x * mv_speed;
 }
